+++ spec/spyke/model/associations_spec.rb
@@ -80,4 +80,74 @@ RSpec.describe FmRest::Spyke::Model::Associations do
       end
     end
   end
+
+  describe "#portals" do
+    it "returns the record's portal associations" do
+      ship = Ship.new
+      expect(ship.portals).to eq([ship.crew, ship.cabins])
+    end
+  end
+
+  describe "#marked_for_destruction" do
+    it "sets the marked_for_destruction flag" do
+      ship = Ship.new
+      expect { ship.mark_for_destruction }.to change { ship.marked_for_destruction? }.from(false).to(true)
+    end
+  end
+
+  describe "#embedded_in_portal" do
+    it "sets the embedded_in_portal flag" do
+      ship = Ship.new
+      expect { ship.embedded_in_portal }.to change { ship.embedded_in_portal? }.from(false).to(true)
+    end
+  end
+
+  describe "#changed? with portal awareness" do
+    it "is true if embedded_in_portal? is true" do
+      ship = Ship.new
+      expect { ship.embedded_in_portal }.to change { ship.changed? }.from(false).to(true)
+    end
+  end
+
+  describe "#changes_applied with portal awareness" do
+    it "resets embedded_in_portal and marked_for_destruction flags" do
+      ship = Ship.new
+      ship.embedded_in_portal
+      ship.mark_for_destruction
+      ship.changes_applied
+      expect(ship.embedded_in_portal?).to eq(false)
+      expect(ship.marked_for_destruction?).to eq(false)
+    end
+  end
+
+  describe "#__new_portal_record_info=" do
+    it "sequentially applies record_ids and mod_ids to new portal records" do
+      ship = Ship.new
+      ship.crew.build(name: "Luffy")
+      ship.crew.build(name: "Hook")
+
+      ship.__new_portal_record_info = [tableName: "PiratesTable", recordId: 5]
+
+      expect(ship.crew[0].__record_id).to eq(4)
+      expect(ship.crew[1].__record_id).to eq(5)
+    end
+  end
+
+  describe "after save: remove_marked_for_destruction" do
+    it "removes portal records marked for destruction" do
+      stub_session_login
+      stub_request(:post, fm_url(layout: "Ships") + "/records").to_return_fm
+
+      ship = Ship.new
+      ship.crew.build(name: "Luffy")
+      ship.crew.build(name: "Hook")
+
+      ship.crew.first.mark_for_destruction
+
+      ship.save
+
+      expect(ship.crew.count).to eq(1)
+      expect(ship.crew.first.name).to eq("Hook")
+    end
+  end
 end